# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.AddCartItemRequest:
    properties:
      productId:
        type: string
      quantity:
        type: integer
    type: object
...
  request.PlaceOrderRequest:
    properties:
      cartId:
        type: string
      paymentMethod:
        type: string
      shippingAddressId:
        type: string
    type: object
...
  request.UpdateCartItemRequest:
    properties:
      quantity:
        type: integer
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      productId:
        type: string
      retryAfter:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
      success:
        type: boolean
    type: object
  response.OrderItemResponse:
    properties:
      productId:
        type: string
      productName:
        type: string
      quantity:
        type: integer
      subtotal:
        type: number
      unitPrice:
        type: number
    type: object
  response.OrderListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/response.OrderResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.OrderResponse:
    properties:
      createdAt:
        type: string
      customerId:
        type: string
      items:
        items:
          $ref: '#/definitions/response.OrderItemResponse'
        type: array
      orderId:
        type: string
      shippingAddressId:
        type: string
      status:
        type: string
      totalAmount:
        type: number
      updatedAt:
        type: string
    type: object
...
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: CS362 Team
  description: Flash Sale E-commerce API with Rate Limiting — CS362 Final Project
  termsOfService: http://swagger.io/terms/
  title: CS362 E-commerce Order System API
  version: "1.0"
paths:
  /cart:
    get:
      description: Retrieves the active shopping cart for the authenticated customer
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Get customer's cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Adds a product to the customer's cart. Increments quantity if product
        already exists.
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Product and quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AddCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/items/{productId}:
    delete:
      description: Removes a specific product from the customer's cart
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Remove item from cart
      tags:
      - Cart
    patch:
      consumes:
      - application/json
      description: Updates the quantity of a specific product in the cart. Set quantity
        to 0 to remove.
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: New quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Update item quantity in cart
      tags:
      - Cart
  /notifications:
    get:
      description: Returns all notifications for the authenticated customer
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Get notifications for a customer
      tags:
      - Notifications
  /notifications/{notificationId}/read:
    patch:
      description: Marks a specific notification as read for the customer
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Notification ID
        in: path
        name: notificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Mark notification as read
      tags:
      - Notifications
  /orders:
    get:
      description: Returns paginated order history for the authenticated customer
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.OrderListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: List orders for a customer
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Creates an order from the customer's cart. Enforces rate limiting
        for flash sale. Returns 429 if rate limit exceeded, 409 if stock is insufficient.
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Order request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.PlaceOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: INSUFFICIENT_STOCK
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "429":
          description: RATE_LIMIT_EXCEEDED
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Place a new order (Flash Sale)
      tags:
      - Orders
  /orders/{orderId}:
    get:
      description: Retrieves a specific order belonging to the authenticated customer
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.OrderResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Get order by ID
      tags:
      - Orders
  /orders/{orderId}/cancel:
    patch:
      description: Cancels an order if it is in PENDING or CONFIRMED state
      parameters:
      - description: Customer ID
        in: header
        name: X-Customer-ID
        required: true
        type: string
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.OrderResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Order cannot be cancelled
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CustomerID: []
      summary: Cancel an order
      tags:
      - Orders
...
securityDefinitions:
  CustomerID:
    description: Customer identifier passed as request header
    in: header
    name: X-Customer-ID
    type: apiKey
swagger: "2.0"